feat: update API documentation to English, refactor validation middleware, and enhance response messages in contact and user routes
