+++ routes/userRoutes.js
@@ -1,16 +1,23 @@
 import express from "express";
 import UserController from "../controllers/userController.js";
-import { UserValidation } from "../controllers/coreController.js";
+import { UserValidation } from "../middleware/userValidation.js";
 
 const userController = new UserController();
 
 const router = express.Router();
+/**
+ * @swagger
+ * tags:
+ *   name: Auth
+ *   description: Endpoints for user authentication and authorization
+ */
+
 
 /**
  * @swagger
  * /auth/register:
  *   post:
- *     summary: Registra um novo usuário
+ *     summary: Register a new user
  *     tags: [Auth]
  *     requestBody:
  *       required: true
@@ -27,9 +34,9 @@ const router = express.Router();
  *                 type: string
  *     responses:
  *       201:
- *         description: Usuário registrado com sucesso
+ *         description: User successfully registered
  *       400:
- *         description: Erro de validação
+ *         description: Validation error
  */
 
 //Router para register a new user
@@ -38,7 +45,7 @@ router.post("/register", UserValidation, userController.register);
  * @swagger
  * /auth/login:
  *   post:
- *     summary: Autentica um usuario
+ *     summary: User authentication
  *     tags: [Auth]
  *     requestBody:
  *       required: true
@@ -53,9 +60,9 @@ router.post("/register", UserValidation, userController.register);
  *                 type: string
  *     responses:
  *       201:
- *         description: Usuário registrado com sucesso
+ *         description: User authenticated successfully
  *       400:
- *         description: Erro de validação
+ *         description: Authentication error
  */
 
 //Router to login the user
